react_django/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
cffi==1.12.3
colorama==0.4.1
cryptography==2.7
Django==2.2.3
django-rest-knox==4.1.0
djangorestframework==3.9.4
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysqlclient==1.4.2
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
